add detail/exceptions.hpp
