+++ Makefile
@@ -5,7 +5,9 @@ SRCS = ${SRCDIR}/json.hpp \
 			 ${SRCDIR}/json_fwd.hpp \
 			 ${SRCDIR}/detail/macro_scope.hpp \
 			 ${SRCDIR}/detail/macro_unscope.hpp \
-			 ${SRCDIR}/detail/meta.hpp
+			 ${SRCDIR}/detail/meta.hpp \
+			 ${SRCDIR}/detail/exceptions.hpp
+
 
 
 # main target
